Sam Commit #4 - Fixed some tests after we change the point struct from int to double. Fixed the data type of the traingle array & the analyze triangle.

+++ PolygonChecker/triangleSolver.c
@@ -4,7 +4,7 @@
 
 #include "triangleSolver.h"
 
-char* analyzeTriangle(int side1, int side2, int side3) {
+char* analyzeTriangle(double side1, double side2, double side3) {
 	char* result = "Not a Triangle";
 
 	if (side1 <= 0 || side2 <= 0 || side3 <= 0 ||
